+++ src/presenter/films-list.js
@@ -4,9 +4,8 @@ import ExtraFilmsListView from "../view/extra-films-list.js";
 import ShowMoreButtonView from "../view/show-more-button.js";
 import NoFilmsView from "../view/no-films.js";
 import LoadingBarView from "../view/loading-bar.js";
-import {render, remove} from "../utils/render.js";
+import {render, remove, getTopCommentedFilms, getTopRatedFilms} from "../utils/render.js";
 import {UpdateType} from "../consts.js";
-import {getTopCommentedFilms, getTopRatedFilms} from "../mock/extra-films";
 import {sortFilmsByRating, sortFilmsByDate} from "../utils/film.js";
 import {SortType} from '../consts.js';
 import FilmCardPresenter from "./film-card";
