+++ src/utils/render.js
@@ -59,4 +59,19 @@ const replace = (newChild, oldChild) => {
   parent.replaceChild(newChild, oldChild);
 };
 
-export {render, RenderPosition, createElement, renderTemplate, remove, replace};
+const START_INDEX = 0;
+const ITEMS_QUANTITY = 2;
+
+const getTopCommentedFilms = (films) => {
+  const sortedFilmsByComments = films.slice().sort((firstItem, secondItem) => secondItem.commentsIds.length - firstItem.commentsIds.length);
+  const topCommented = sortedFilmsByComments.filter((film) => film.commentsIds.length > 0);
+  return topCommented.splice(START_INDEX, ITEMS_QUANTITY);
+};
+
+const getTopRatedFilms = (films) => {
+  const sortedFilmsByComments = films.slice().sort((firstItem, secondItem) => secondItem.rating - firstItem.rating);
+  const topRated = sortedFilmsByComments.filter((film) => film.rating > 0);
+  return topRated.splice(START_INDEX, ITEMS_QUANTITY);
+};
+
+export {render, RenderPosition, createElement, renderTemplate, remove, replace, getTopCommentedFilms, getTopRatedFilms};
